backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.2.0
pyodbc>=5.1.0
httpx>=0.27.0
python-multipart>=0.0.9
